Extract auto-complete logic into testable service module
Move checkAutoComplete() and showAutoCompleteNotification() from inline
extension.ts closure into packages/core/src/services/autoComplete.ts.
Enables unit testing without mocking the full extension wiring.
